Use rtree for creating mosaic
