+++ cogeo_mosaic/utils.py
@@ -192,129 +192,6 @@ def _filter_and_sort(tile, dataset, minimum_cover=None, sort_cover=False):
     return [x[1] for x in dataset]
 
 
-def create_mosaic(
-    dataset_list: Tuple,
-    minzoom: int = None,
-    maxzoom: int = None,
-    max_threads: int = 20,
-    minimum_tile_cover: float = None,
-    tile_cover_sort: bool = False,
-    version: str = "0.0.2",
-    quiet: bool = True,
-) -> Dict:
-    """
-    Create mosaic definition content.
-
-    Attributes
-    ----------
-        dataset_list : tuple or list, required
-            Dataset urls.
-        minzoom: int, optional
-            Force mosaic min-zoom.
-        maxzoom: int, optional
-            Force mosaic max-zoom.
-        minimum_tile_cover: float, optional (default: 0)
-            Filter files with low tile intersection coverage.
-        tile_cover_sort: bool, optional (default: None)
-            Sort intersecting files by coverage.
-        max_threads : int
-            Max threads to use (default: 20).
-        version: str, optional
-            mosaicJSON definition version
-        quiet: bool, optional (default: True)
-            Mask processing steps.
-
-    Returns
-    -------
-        mosaic_definition : dict
-            Mosaic definition.
-
-    """
-    if version not in ["0.0.1", "0.0.2"]:
-        raise Exception(f"Invalid mosaicJSON's version: {version}")
-
-    if not quiet:
-        click.echo("Get files footprint", err=True)
-
-    results = get_footprints(dataset_list, max_threads=max_threads, quiet=quiet)
-
-    if minzoom is None:
-        minzoom = list(set([feat["properties"]["minzoom"] for feat in results]))
-        if len(minzoom) > 1:
-            warnings.warn(
-                "Multiple MinZoom, Assets different minzoom values", UserWarning
-            )
-
-        minzoom = max(minzoom)
-
-    if maxzoom is None:
-        maxzoom = list(set([feat["properties"]["maxzoom"] for feat in results]))
-        if len(maxzoom) > 1:
-            warnings.warn(
-                "Multiple MaxZoom, Assets have multiple resolution values", UserWarning
-            )
-
-        maxzoom = max(maxzoom)
-
-    quadkey_zoom = minzoom
-
-    datatype = list(set([feat["properties"]["datatype"] for feat in results]))
-    if len(datatype) > 1:
-        raise Exception("Dataset should have the same data type")
-
-    if not quiet:
-        click.echo(f"Get quadkey list for zoom: {quadkey_zoom}", err=True)
-
-    tiles = burntiles.burn(results, quadkey_zoom)
-    tiles = ["{2}-{0}-{1}".format(*tile.tolist()) for tile in tiles]
-
-    bounds = burntiles.find_extrema(results)
-    mosaic_definition = dict(
-        mosaicjson=version,
-        minzoom=minzoom,
-        maxzoom=maxzoom,
-        bounds=bounds,
-        center=[(bounds[0] + bounds[2]) / 2, (bounds[1] + bounds[3]) / 2, minzoom],
-        tiles={},
-        version="1.0.0",
-    )
-
-    if version == "0.0.2":
-        mosaic_definition.update(dict(quadkey_zoom=quadkey_zoom))
-
-    if not quiet:
-        click.echo(f"Feed Quadkey index", err=True)
-
-    dataset_geoms = polygons([feat["geometry"]["coordinates"][0] for feat in results])
-    dataset = [
-        {"path": f["properties"]["path"], "geometry": geom}
-        for (f, geom) in zip(results, dataset_geoms)
-    ]
-
-    for parent in tiles:
-        z, x, y = list(map(int, parent.split("-")))
-        parent = mercantile.Tile(x=x, y=y, z=z)
-        quad = mercantile.quadkey(*parent)
-        tile_geometry = polygons(
-            mercantile.feature(parent)["geometry"]["coordinates"][0]
-        )
-        fdataset = [
-            dataset[idx]
-            for idx in numpy.nonzero(intersects(tile_geometry, dataset_geoms))[0]
-        ]
-        if minimum_tile_cover is not None or tile_cover_sort:
-            fdataset = _filter_and_sort(
-                tile_geometry,
-                fdataset,
-                minimum_cover=minimum_tile_cover,
-                sort_cover=tile_cover_sort,
-            )
-        if len(fdataset):
-            mosaic_definition["tiles"][quad] = [f["path"] for f in fdataset]
-
-    return mosaic_definition
-
-
 def update_mosaic(
     dataset_list: Tuple,
     mosaic_def: dict,
